feat: enhance video upload validation and error handling in video controller

+++ src/utils/cloudinary.js
@@ -19,7 +19,9 @@ const uploadOnCloudinary = async (localFilePath) => {
     fs.unlinkSync(localFilePath);
     return response;
   } catch (error) {
-    fs.unlinkSync(localFilePath); // remove the locally saved temporary files as the upload operation got failed
+    if (fs.existsSync(localFilePath)) {
+      fs.unlinkSync(localFilePath); // remove the locally saved temporary files as the upload operation got failed
+    }
     return null;
   }
 };
